requirements.txt
streamlit==1.39.0
pybaseball==2.2.7
pandas==2.2.3
altair==5.4.1
streamlit-searchbox==0.1.16
numpy==1.26.4
scipy==1.14.1
matplotlib==3.10.6
Pillow==11.3.0
requests==2.32.5